Inversed if mod exists logic
Previously it was decrease opacity if count was = 0, but when it comes to duplicated images, sometimes it will be of lower opacity even if the count is > 0. Current logic is if count > 0, then it is opaque. This logic makes more sense too as it should all start out as 0, then find for matches then populate the data.